+++ crates/rpc-types-txpool/src/txpool.rs
@@ -16,7 +16,7 @@ pub struct TxpoolInspectSummary {
     /// Transferred value
     pub value: U256,
     /// Gas amount
-    pub gas: u128,
+    pub gas: u64,
     /// Gas Price
     pub gas_price: u128,
 }
@@ -60,7 +60,7 @@ impl Visitor<'_> for TxpoolInspectSummaryVisitor {
             }
         };
         let value = U256::from_str(value_split[0]).map_err(de::Error::custom)?;
-        let gas = u128::from_str(gas_split[0]).map_err(de::Error::custom)?;
+        let gas = u64::from_str(gas_split[0]).map_err(de::Error::custom)?;
         let gas_price = u128::from_str(gas_price_split[0]).map_err(de::Error::custom)?;
 
         Ok(TxpoolInspectSummary { to, value, gas, gas_price })
@@ -423,7 +423,7 @@ mod tests {
             TxpoolInspectSummary {
                 to: Some(Address::from_str("000000000000000000000000000000000000007E").unwrap()),
                 value: U256::from(0u128),
-                gas: 100187u128,
+                gas: 100187,
                 gas_price: 20000000000u128,
             },
         );
@@ -437,7 +437,7 @@ mod tests {
             TxpoolInspectSummary {
                 to: Some(Address::from_str("d10e3Be2bc8f959Bc8C41CF65F60dE721cF89ADF").unwrap()),
                 value: U256::from(0u128),
-                gas: 65792u128,
+                gas: 65792,
                 gas_price: 2000000000u128,
             },
         );
@@ -446,7 +446,7 @@ mod tests {
             TxpoolInspectSummary {
                 to: Some(Address::from_str("d10e3Be2bc8f959Bc8C41CF65F60dE721cF89ADF").unwrap()),
                 value: U256::from(0u128),
-                gas: 65792u128,
+                gas: 65792,
                 gas_price: 2000000000u128,
             },
         );
@@ -455,7 +455,7 @@ mod tests {
             TxpoolInspectSummary {
                 to: Some(Address::from_str("d10e3Be2bc8f959Bc8C41CF65F60dE721cF89ADF").unwrap()),
                 value: U256::from(0u128),
-                gas: 65780u128,
+                gas: 65780,
                 gas_price: 2000000000u128,
             },
         );
@@ -464,7 +464,7 @@ mod tests {
             TxpoolInspectSummary {
                 to: Some(Address::from_str("d10e3Be2bc8f959Bc8C41CF65F60dE721cF89ADF").unwrap()),
                 value: U256::from(0u128),
-                gas: 65780u128,
+                gas: 65780,
                 gas_price: 2000000000u128,
             },
         );
@@ -478,7 +478,7 @@ mod tests {
             TxpoolInspectSummary {
                 to: None,
                 value: U256::from(0u128),
-                gas: 612412u128,
+                gas: 612412,
                 gas_price: 6000000000u128,
             },
         );
@@ -493,7 +493,7 @@ mod tests {
             TxpoolInspectSummary {
                 to: Some(Address::from_str("3479BE69e07E838D9738a301Bb0c89e8EA2Bef4a").unwrap()),
                 value: U256::from(1000000000000000u128),
-                gas: 21000u128,
+                gas: 21000,
                 gas_price: 10000000000u128,
             },
         );
@@ -502,7 +502,7 @@ mod tests {
             TxpoolInspectSummary {
                 to: Some(Address::from_str("73Aaf691bc33fe38f86260338EF88f9897eCaa4F").unwrap()),
                 value: U256::from(1000000000000000u128),
-                gas: 21000u128,
+                gas: 21000,
                 gas_price: 10000000000u128,
             },
         );
@@ -516,7 +516,7 @@ mod tests {
             TxpoolInspectSummary {
                 to: Some(Address::from_str("73Aaf691bc33fe38f86260338EF88f9897eCaa4F").unwrap()),
                 value: U256::from(10000000000000000u128),
-                gas: 21000u128,
+                gas: 21000,
                 gas_price: 10000000000u128,
             },
         );
